@@ -1,74 +1,37 @@
 {
     "name": "laravel/laravel",
     "type": "project",
-    "description": "The Laravel Framework.",
-    "keywords": ["framework", "laravel"],
-    "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0",
-        "ext-json": "*",
-        "doctrine/dbal": "~2.3",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
-        "infyomlabs/adminlte-templates": "^3.0",
-        "infyomlabs/generator-builder": "^1.0",
-        "infyomlabs/laravel-generator": "^3.0",
-        "laravel/framework": "^8.75",
-        "laravel/sanctum": "^2.11",
-        "laravel/tinker": "^2.5",
-        "laravel/ui": "^3.4",
-        "laravelista/comments": "^4.5",
-        "livewire/livewire": "^2.12",
-        "onelogin/php-saml": "*",
-        "spatie/laravel-medialibrary": "^9.12",
-        "yajra/laravel-datatables-buttons": "*",
-        "yajra/laravel-datatables-html": "*",
-        "yajra/laravel-datatables-oracle": "*"
+        "php": "^8.1",
+        "lakm/laravel-comments": "^2.2",
+        "laravel/framework": "^11.0",
+        "laravel/tinker": "^2.10",
+        "laravel/ui": "^4.6",
+        "livewire/livewire": "^3.0",
+        "masmerise/livewire-toaster": "^2.6",
+        "power-components/livewire-powergrid": "^6.1.11",
+        "spatie/laravel-medialibrary": "^11.0"
     },
     "require-dev": {
-        "facade/ignition": "^2.5",
         "fakerphp/faker": "^1.9.1",
-        "laravel/sail": "^1.0.1",
+        "laravel/breeze": "^2.3",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^5.10",
-        "phpunit/phpunit": "^9.5.10"
+        "pestphp/pest": "^3.7",
+        "pestphp/pest-plugin-laravel": "^3.0",
+        "spatie/laravel-ignition": "^2.9"
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/",
-            "Database\\Factories\\": "database/factories/",
-            "Database\\Seeders\\": "database/seeders/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Tests\\": "tests/"
-        }
-    },
-    "scripts": {
-        "post-autoload-dump": [
-            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
-            "@php artisan package:discover --ansi"
-        ],
-        "post-update-cmd": [
-            "@php artisan vendor:publish --tag=laravel-assets --ansi --force"
-        ],
-        "post-root-package-install": [
-            "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
-        ],
-        "post-create-project-cmd": [
-            "@php artisan key:generate --ansi"
-        ]
-    },
-    "extra": {
-        "laravel": {
-            "dont-discover": []
+            "App\\": "app/"
         }
     },
     "config": {
         "optimize-autoloader": true,
         "preferred-install": "dist",
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "pestphp/pest-plugin": true
+        }
     },
     "minimum-stability": "dev",
     "prefer-stable": true
