@@ -3,19 +3,13 @@
 namespace App\Repositories;
 
 use App\Models\User;
-use App\Repositories\BaseRepository;
-
-/**
- * Class UserRepository
- * @package App\Repositories
-*/
 
 class UserRepository extends BaseRepository
 {
     /**
      * @var array
      */
-    protected $fieldSearchable = [
+    protected array $fieldSearchable = [
         'name',
         'email',
         'password'
@@ -26,15 +20,17 @@ class UserRepository extends BaseRepository
      *
      * @return array
      */
-    public function getFieldsSearchable()
+    public function getFieldsSearchable(): array
     {
         return $this->fieldSearchable;
     }
 
     /**
      * Configure the Model
-     **/
-    public function model()
+     *
+     * @return string
+     */
+    public function model(): string
     {
         return User::class;
     }
