@@ -1,1 +1,11 @@
-require('./bootstrap');
+console.log('App.js loaded');
+import './bootstrap';
+
+// PowerGrid assets
+import '../../vendor/power-components/livewire-powergrid/dist/powergrid'
+import '../../vendor/power-components/livewire-powergrid/dist/tailwind.css'
+
+// resources/js/app.js
+
+import flatpickr from "flatpickr";
+import 'flatpickr/dist/flatpickr.min.css'
