@@ -2,16 +2,15 @@
 
 @section('content')
     <x-page-header title="Campuses" text="Edit and manage campuses, and assign librarians to notify when new requests arrive.">
-{{--        <a class="btn btn-primary float-right"--}}
-{{--           href="{{ route('campuses.create') }}">--}}
+{{--        <a class="btn btn-primary float-right" href="{{ route('campuses.create') }}">--}}
 {{--            Add New--}}
 {{--        </a>--}}
     </x-page-header>
 
-
     <div class="content px-3">
 
-        @include('flash::message')
+        {{-- Display flash messages --}}
+{{--        @include('flash::message')--}}
 
         <div class="clearfix"></div>
 
@@ -19,9 +18,6 @@
             <div class="card-body p-0">
                 @include('campuses.table')
             </div>
-
         </div>
     </div>
-
 @endsection
-
