@@ -3,15 +3,15 @@
 @section('content')
 
     <x-page-header title="Classes" text="Edit and manage classes.">
-            <a class="btn btn-primary float-right"
-               href="{{ route('classes.create') }}">
-                Add New
-            </a>
+{{--            <a class="btn btn-primary float-right"--}}
+{{--               href="{{ route('classes.create') }}">--}}
+{{--                Add New--}}
+{{--            </a>--}}
     </x-page-header>
 
     <div class="content px-3">
 
-        @include('flash::message')
+{{--        @include('flash::message')--}}
 
         <div class="clearfix"></div>
 
