@@ -89,7 +89,7 @@
 />
 
 <!-- Class Notes (Textarea) -->
-<x-textarea
+<x-input-textarea
     name="class_notes"
     label="Class Notes"
     :value="$instructionRequestDetails->class_notes ?? null"
@@ -103,7 +103,7 @@
 />
 
 <!-- Assessment Notes (Textarea) -->
-<x-textarea
+<x-input-textarea
     name="assessment_notes"
     label="Assessment Notes"
     :value="$instructionRequestDetails->assessment_notes ?? null"
